fix the dependency of the position of the camera

+++ src/geometry/vec3d.h
@@ -5,7 +5,7 @@
 
 namespace geometry {
 
-using dim_t = float;
+using dim_t = double;
 
 struct vec3d {
     dim_t x, y, z;
